requirements.txt
# LLM + embeddings + TTS
openai>=1.0.0

# Gradio (Hugging Face Spaces runtime)
gradio>=4.44.0

# Vector database
chromadb>=0.5.0

# Document parsing
pdfplumber>=0.9.0
python-pptx>=0.6.21
trafilatura>=1.6.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0